chore: move randomizeArray from numbers module to arrays module
